requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
captcha==0.3
certifi==2021.10.8
charset-normalizer==2.0.9
Deprecated==1.2.13
Django==3.2.7
django-redis==5.2.0
djangorestframework==3.13.1
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.4.0
PyMySQL==1.0.2
pyparsing==3.0.6
pytz==2021.3
redis==4.1.0
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.7
wrapt==1.13.3
